requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
alembic==1.2.1
antlr4-python3-runtime==4.8
async-timeout==3.0.1
attrs==21.2.0
ax-platform==0.2.1
botorch==0.5.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
cloudpickle==1.6.0
configparser==5.0.2
databricks-cli==0.9.0
docker==5.0.0
entrypoints==0.3
Flask==2.0.1
fsspec==2021.7.0
future==0.18.2
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.35.0
google-auth-oauthlib==0.4.5
gpytorch==1.5.0
greenlet==1.1.0
grpcio==1.38.1
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
Mako==1.1.5
Markdown==3.3.4
MarkupSafe==2.0.1
mlflow==1.19.0
multidict==5.1.0
numpy==1.21.0
oauthlib==3.1.1
omegaconf==2.1.1
packaging==21.0
pandas==1.3.0
Pillow==8.3.1
plotly==5.2.2
prometheus-client==0.11.0
prometheus-flask-exporter==0.15.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-editor==1.0.4
pytorch-lightning==1.4.2
pytz==2021.1
PyYAML==5.4.1
querystring-parser==1.2.4
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.7.0
six==1.16.0
smmap==4.0.0
SQLAlchemy==1.4.21
sqlparse==0.4.1
tabulate==0.8.9
tenacity==8.0.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
threadpoolctl==2.2.0
torch==1.10.0.dev20210823+cu102
torchaudio==0.10.0.dev20210823
torchmetrics==0.5.0
torchvision==0.11.0.dev20210823+cu102
tqdm==4.62.2
typeguard==2.12.1
typing-extensions==3.10.0.0
urllib3==1.26.6
websocket-client==1.2.1
Werkzeug==2.0.1
yarl==1.6.3